+++ ods_tools/oed/common.py
@@ -190,6 +190,7 @@ DEFAULT_VALIDATION_CONFIG = [
     {'name': 'construction_code', 'on_error': 'raise'},
     {'name': 'country_and_area_code', 'on_error': 'raise'},
     {'name': 'conditional_requirement', 'on_error': 'raise'},
+    {'name': 'oedversion_consistency', 'on_error': 'raise'},
 ]
 
 OED_PERIL_COLUMNS = ['AccPeril', 'PolPerilsCovered', 'PolPeril', 'CondPeril', 'LocPerilsCovered', 'LocPeril',
@@ -220,6 +221,8 @@ default_string_dtype = {
     'pd_dtype': 'category',
     'pa_dtype': 'string[pyarrow]',
 }
+
+pd_default_string = object if pd.__version__ < "3" else "str"
 
 
 def is_empty(df, columns):